pyproject.toml
[tool.poetry]
name = "django-boilerplate"
version = "0.1.0"
description = ""
authors = ["bucky <buckyroberts@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
django-filter = "^22.1"
django-split-settings = "^1.2.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
